traffic lights cleanup
